Update Final_Final of Finals Finally.s
I changed the start to have the turn dynamics from 3.3 bc it will be very hard to try to change it to not be like that. also put in uppercase comments that I think I can't address wo your code tho I might be able to now? idk my brain is seizing looking at this lol